+++ stable_baselines/common/base_class.py
@@ -312,8 +312,8 @@ class TensorboardWriter:
         max_run_id = 0
         for path in glob.glob(self.tensorboard_log_path + "/{}_[0-9]*".format(self.tb_log_name)):
             file_name = path.split("/")[-1]
-            ext = "_".join(file_name.split("_")[1:])
-            if ext.isdigit() and int(ext) > max_run_id:
+            ext = file_name.split("_")[-1]
+            if self.tb_log_name == "_".join(file_name.split("_")[:-1]) and ext.isdigit() and int(ext) > max_run_id:
                 max_run_id = int(ext)
         return max_run_id