+++ docs/misc/changelog.rst
@@ -21,6 +21,7 @@ Master version 1.0.8.a0 (WIP)
 - fixed image normalization not occurring when using images
 - updated VecEnv objects for the new Gym version
 - changed default policies for DDPG, so that DDPG now works correctly
+- refactored DQN policies
 
 
 Release 1.0.7 (2018-08-29)